implement asset page redesign and add base apr line to graph
