+++ src/components/asset/AssetSupplyInfo.tsx
@@ -1,19 +1,27 @@
 import { Circle } from '@mui/icons-material';
 import { Box, Typography, useTheme } from '@mui/material';
-import { usePool, usePoolOracle } from '../../hooks/api';
+import { useBackstop, usePool, usePoolOracle } from '../../hooks/api';
 import { toBalance, toPercentage } from '../../utils/formatter';
+import { estimateEmissionsApr } from '../../utils/math';
+import { AprDisplay } from '../common/AprDisplay';
 import { ReserveComponentProps } from '../common/ReserveComponentProps';
 import { Row } from '../common/Row';
 import { Section, SectionSize } from '../common/Section';
-import { AssetGraphBorrow } from './AssetGraphBorrow';
 
-export const AssetGraphBoxBorrow: React.FC<ReserveComponentProps> = ({ poolId, assetId }) => {
-  const theme = useTheme();
+export const AssetSupplyInfo: React.FC<ReserveComponentProps> = ({ poolId, assetId }) => {
   const { data: pool } = usePool(poolId);
   const { data: poolOracle } = usePoolOracle(pool);
+  const { data: backstop } = useBackstop();
   const oraclePrice = poolOracle?.getPriceFloat(assetId);
   const reserve = pool?.reserves.get(assetId);
+  const emissionsPerAsset = reserve?.emissionsPerYearPerSuppliedAsset();
+  const emissionApr =
+    backstop && emissionsPerAsset && emissionsPerAsset > 0 && oraclePrice
+      ? estimateEmissionsApr(emissionsPerAsset, backstop.backstopToken, oraclePrice)
+      : undefined;
   const hasData = pool && poolOracle && reserve && oraclePrice;
+  const theme = useTheme();
+
   return (
     <>
       {hasData && (
@@ -24,7 +32,6 @@ export const AssetGraphBoxBorrow: React.FC<ReserveComponentProps> = ({ poolId, a
           <Row sx={{ padding: '6px' }}>
             <Typography
               sx={{
-                width: '100%',
                 padding: '6px',
                 display: 'flex',
                 gap: '4px',
@@ -33,8 +40,8 @@ export const AssetGraphBoxBorrow: React.FC<ReserveComponentProps> = ({ poolId, a
               variant="body2"
               color={theme.palette.text.primary}
             >
-              <Circle fontSize="inherit" sx={{ width: '8px', color: theme.palette.borrow.main }} />
-              Borrow Info
+              <Circle fontSize="inherit" sx={{ width: '8px', color: theme.palette.lend.main }} />
+              Supply Info
             </Typography>
           </Row>
           <Row>
@@ -49,9 +56,32 @@ export const AssetGraphBoxBorrow: React.FC<ReserveComponentProps> = ({ poolId, a
                 background: theme.palette.background.default,
               }}
             >
-              <Typography sx={{ padding: '6px' }}>Total Borrowed</Typography>
-              <Typography sx={{ padding: '6px', color: theme.palette.borrow.main }}>
-                {toBalance(reserve.totalLiabilitiesFloat() * oraclePrice)}
+              <Typography sx={{ padding: '6px' }}>APR</Typography>
+              <AprDisplay
+                assetSymbol={reserve.tokenMetadata.symbol}
+                assetApr={reserve.supplyApr}
+                emissionSymbol={'BLND'}
+                emissionApr={emissionApr}
+                isSupply={true}
+                direction={'horizontal'}
+              />
+            </Box>
+          </Row>
+          <Row>
+            <Box
+              sx={{
+                width: '100%',
+                padding: '6px',
+                margin: '6px',
+                borderRadius: '5px',
+                display: 'flex',
+                justifyContent: 'space-between',
+                background: theme.palette.background.default,
+              }}
+            >
+              <Typography sx={{ padding: '6px' }}>Total Supplied</Typography>
+              <Typography sx={{ padding: '6px', color: theme.palette.lend.main }}>
+                {toBalance(reserve.totalSupplyFloat() * oraclePrice)}
               </Typography>
             </Box>
           </Row>
@@ -67,14 +97,11 @@ export const AssetGraphBoxBorrow: React.FC<ReserveComponentProps> = ({ poolId, a
                 background: theme.palette.background.default,
               }}
             >
-              <Typography sx={{ padding: '6px' }}>Liability Factor</Typography>
+              <Typography sx={{ padding: '6px' }}>Collateral Factor</Typography>
               <Typography sx={{ padding: '6px' }}>
-                {toPercentage(reserve.getLiabilityFactor())}
+                {toPercentage(reserve.getCollateralFactor())}
               </Typography>
             </Box>
-          </Row>
-          <Row>
-            <AssetGraphBorrow poolId={poolId} assetId={assetId} reserve={reserve} />
           </Row>
         </Section>
       )}
